% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myEigen_QR.R
\name{myEigen_QR}
\alias{myEigen_QR}
\title{PCA using QR Decomposition}
\usage{
myEigen_QR(A, numIter = 1000)
}
\arguments{
\item{A}{A square, symmetric matrix}

\item{numIter}{the number of iterations to run QR decomposition on, default is 1000}
}
\value{
Returns a list of eigenvalues and eigenvectors of A
}
\description{
This function performs PCA using QR decomposition.
}
\examples{
set.seed(239)
n <- 100
p <- 3
X <- matrix(rnorm(n * p), nrow = 100)
A <- t(X) \%*\% X

}
