% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLM.R
\name{myLM}
\alias{myLM}
\title{Linear Model Function}
\usage{
myLM(X, Y, intercept = TRUE)
}
\arguments{
\item{X}{an n x p matrix of explanatory variables}

\item{Y}{an n dimensional vector of responses}

\item{intercept}{Boolean for whether an intercept term should be included, default is true}
}
\value{
Returns the OLS betas along with their respective standard errors
}
\description{
This function calculates the regression coefficients for linear regression as well as standard errors
}
\examples{
set.seed(12083)
X <- matrix(rnorm(100), 50, 2)
Y <- matrix(3 * X[,1] + 2 * X[,2] + rnorm(50))
myLM(X, Y)

}
