% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLogistic.R
\name{myLogistic}
\alias{myLogistic}
\title{Logistic Regression Fitting}
\usage{
myLogistic(X, Y)
}
\arguments{
\item{X}{an n x p matrix of explanatory variables}

\item{Y}{an n dimensional vector of binary responses}
}
\value{
A list of the logistic betas and their respective standard errors
}
\description{
This function performs the logistic regression of Y on X and returns the coeffients. No intercept term is assumed
}
\examples{
set.seed(18274)
n <- 500
p <- 4
X    <- matrix(rnorm(n * p), nrow = n)
beta <- c(12, -2,-3, 4)
Y    <- 1 * (runif(n) <  1 / (1 + exp(-(X \%*\% beta))))
logistic_beta <- myLogistic(X, Y)
logistic_beta

}
