% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myqR.R
\name{myQR}
\alias{myQR}
\title{QR Decomposition}
\usage{
myQR(A)
}
\arguments{
\item{A}{an n by m  matrix whose QR decomposition is to be computed}
}
\value{
Q, an orthogonal matrix Q  and an upper triangular matrix R
}
\description{
This function performs QR decomposition on a matrix  A into a product A = QR of an orthogonal matrix Q and an upper triangular matrix R
}
\examples{
A <- rbind(c(-13, 14, -2), c(0, 2, 2), c(3, 9, -7))
myQR(A)
}
