% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ridge.R
\name{myRidge}
\alias{myRidge}
\title{Ridge Regression}
\usage{
myRidge(X, Y, lambda)
}
\arguments{
\item{X}{an n x p matrix of explanatory variables.}

\item{Y}{an n vector or matrix of dependent variables.}

\item{lambda}{regularization parameter for lambda >= 0}
}
\value{
Returns a vector of betas, the ridge regression solution with p + 1 elements
}
\description{
This function performs ridge regression of Y on X.
}
\examples{
set.seed(12837)
X = runif(50)
X = matrix(sort(X), nrow = 50)
sigma = .1
Y =X^2 + rnorm(50)*sigma
lambda = 0
beta_r <- myRidge(X, Y, lambda)
Yhat <- cbind(rep(1, length(X)), X)\%*\%beta_r
plot(X, Y, col = 'indianred')
par(new =TRUE
plot(X, Yhat, type = 'l', col='slateblue')
abline(coef(lm(Y~X)))
}
